Movie scoring updates.
